fix: respond to pr comments
