+++ app/crud/word_of_day_repository.py
@@ -15,5 +15,5 @@ class WordOfDayRepository(GenericRepository[WordOfDay]):
         return self._session.exec(stmt).one()
 
     def add_word_of_day(self, word_id: int) -> None:
-        word_of_day = WordOfDay(word_id=word_id)
+        word_of_day = WordOfDay(word_id=word_id, date=date.today())
         self.upsert(word_of_day)
